Use new page action icons and title for validation results
This implements issue #2 and shows the final validation status in the page action icon as OK, WARNING or ERROR. Additionally the summary status string is set as title for the icon. Therefore quick-checking the validation status of a page should become easier. The svg source versions of icons have also been moved to a new directory.